requirements-ci.txt
alabaster==0.7.10
appdirs==1.4.3
asn1crypto==0.22.0
astroid==1.5.3
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
Babel==2.4.0
backports.functools-lru-cache==1.4
blockdiag==1.5.3
boto==2.48.0
boto3==1.4.4
botocore==1.5.90
cffi==1.10.0
click==6.7
configparser==3.5.0
constantly==15.1.0
cookies==2.2.1
cryptography==2.0.2
decorator==4.1.2
dicttoxml==1.7.4
docutils==0.13.1
enum34==1.1.6
flake8==3.4.1
funcparserlib==0.3.6
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
idna==2.5
imagesize==0.7.1
incremental==17.5.0
ipaddress==1.0.18
isort==4.2.15
Jinja2==2.9.6
jmespath==0.9.3
jsonref==0.1
lazy-object-proxy==1.3.1
lz4==0.10.1
markdown2==2.3.4
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
moto==1.0.1
olefile==0.44
packaging==16.8
pathlib2==2.3.0
pbr==3.1.1
pep8==1.7.0
Pillow==4.2.1
pyaml==17.7.2
pyasn1==0.3.1
pyasn1-modules==0.0.10
pycodestyle==2.3.1
pycparser==2.18
pyenchant==1.6.11
pyflakes==1.5.0
pyjade==4.0.0
PyJWT==1.5.2
pylint==1.7.2
pyOpenSSL==17.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
ramlfications==0.1.9
requests==2.18.2
s3transfer==0.1.10
scandir==1.5
service-identity==17.0.0
setuptools-trial==0.6.0
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
SQLAlchemy==1.1.12
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
termcolor==1.1.0
toml==0.9.2
towncrier==17.4.0
treq==17.3.1
Twisted==17.5.0
txaio==2.8.1
txgithub==15.0.0
txrequests==0.9.5
typing==3.6.1
webcolors==1.7
Werkzeug==0.12.2
wrapt==1.10.10
xmltodict==0.11.0
zope.interface==4.4.2
coverage==4.4.1
codecov==2.0.9
-e master
-e worker
-e pkg
# we install buildbot www from pypi to avoid the slow nodejs build at each test
buildbot-www==0.9.9.post2
